requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.12
dj-database-url==2.1.0
drf-spectacular==0.27.0
Pillow==12.0.0
reportlab==4.0.9
python-decouple==3.8
django-filter==23.5
python-dateutil==2.8.2